Extract highlighted text matching helpers
