+++ src/lib/string/index.ts
@@ -1,5 +1,6 @@
 export * from "./capitalize";
 export * from "./escapeRegExp";
+export * from "./getHighlightedTextSegments";
 export * from "./isValidSlug";
 export * from "./pluralize";
 export * from "./sanitizeInput";
